==> MPP_DS/2. Prepared Test Data.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns
import scipy.stats as ss
import math



from sklearn import preprocessing
import sklearn.model_selection as ms
from sklearn import linear_model
import sklearn.metrics as sklm
from sklearn import feature_selection as fs
from sklearn import metrics
from sklearn.model_selection import cross_validate

%matplotlib inline

pd.set_option('display.float_format', lambda x: '{:.3f}'.format(x)) #Limiting floats output to 3 decimal points

import warnings
warnings.filterwarnings('ignore')

df_test = pd.read_csv('test_values.csv')


print(df_test['rucc'].unique())
Features = df_test['rucc']
enc = preprocessing.LabelEncoder()
enc.fit(Features)
Features = enc.transform(Features)
print(Features)
ohe = preprocessing.OneHotEncoder()
encoded = ohe.fit(Features.reshape(-1,1))
Features = encoded.transform(Features.reshape(-1,1)).toarray()
Features[:10,:]

def encode_string(cat_feature):
    ## First encode the strings to numeric categories
    enc = preprocessing.LabelEncoder()
    enc.fit(cat_feature)
    enc_cat_feature = enc.transform(cat_feature)
    ## Now, apply one hot encoding
    ohe = preprocessing.OneHotEncoder()
    encoded = ohe.fit(enc_cat_feature.reshape(-1,1))
    return encoded.transform(enc_cat_feature.reshape(-1,1)).toarray()
    
categorical_columns= ['economic_typology','urban_influence','year']

for col in categorical_columns:
    temp = encode_string(df_test[col])
    Features = np.concatenate([Features, temp], axis = 1)

print(Features.shape)
print(Features[:2, :]) 

num_cols=['renter_occupied_households',
 'pct_renter_occupied',
 'median_gross_rent',
 'median_household_income',
 'median_property_value',
 'rent_burden',
 'pct_white',
 'pct_af_am',
 'pct_hispanic',
 'pct_am_ind',
 'pct_asian',
 'pct_nh_pi',
 'pct_multiple',
 'pct_other',
 'poverty_rate',
 'pct_civilian_labor',
 'pct_unemployment',
 'pct_uninsured_adults',
 'pct_uninsured_children',
 'pct_adult_obesity',
 'pct_diabetes',
 'pct_physical_inactivity',
 'heart_disease_mortality_per_100k',
 'pct_female',
 'pct_below_18_years_of_age',
 'pct_aged_65_years_and_older',
 'pct_adults_less_than_a_high_school_diploma',
 'pct_adults_with_high_school_diploma',
 'pct_adults_with_some_college',
 'pct_adults_bachelors_or_higher',
 'birth_rate_per_1k',
 'death_rate_per_1k']

enc_cols = [
    'RUCC1', 'RUCC2','RUCC3','RUCC4','RUCC5','RUCC6','RUCC7','RUCC8','RUCC9',
    'ECT1','ECT2','ECT3','ECT4','ECT5','ECT6',
    'UI1','UI2','UI3','UI4','UI5','UI6','UI7','UI8','UI9','UI10','UI11','UI12',
    'YA','YB']

cols = num_cols+enc_cols

df_test= df_test.fillna(df_test.mean())

df_test_enc= np.concatenate([df_test[num_cols],Features],axis=1)
df_test_enc = pd.DataFrame(df_test_enc, columns=cols)
df_test_enc.to_csv('df_test_enc.csv')

df_test_enc.shape

## transforming numerical data

data_norm = preprocessing.normalize(df_test_enc[num_cols], axis = 0)

df_test_trf = np.concatenate([data_norm,Features],axis=1)

df_test_trf = pd.DataFrame(df_test_trf, columns=cols)
df_test_trf.to_csv('df_test_trf.csv')

df_test_trf.shape
