==> MPP_DS/3c. Stacked Model.ipynb <==
#Fundamental librarys to math and stats process
import numpy as np
import numpy.random as nr
import scipy.stats as ss
import math
#data prepared
import pandas as pd

#ML preprocessi
from sklearn import preprocessing
from sklearn.preprocessing import StandardScaler
import sklearn.model_selection as ms
from sklearn.pipeline import make_pipeline
from sklearn import feature_selection as fs

# ML algorithms models
from sklearn.ensemble import RandomForestRegressor,GradientBoostingRegressor 
from sklearn import linear_model as lm

# ML Evaluations
import sklearn.metrics as sklm
from sklearn import metrics
from sklearn.model_selection import cross_validate, train_test_split

#Ploting
import matplotlib.pyplot as plt
import seaborn as sns


%matplotlib inline

df = (pd.read_csv('x.csv'))
y = (pd.read_csv('y.csv'))
y = y['evictions']
df = df.drop(['Unnamed: 0'], axis = 1)
print(df.shape)
print(y.shape)

list(df.columns)

df.head()

### Encode categorical featured

print(df['rucc'].unique())
Features = df['rucc']
enc = preprocessing.LabelEncoder()
enc.fit(Features)
Features = enc.transform(Features)
print(Features)

ohe = preprocessing.OneHotEncoder()
encoded = ohe.fit(Features.reshape(-1,1))
Features = encoded.transform(Features.reshape(-1,1)).toarray()
Features[:10,:]

def encode_string(cat_feature):
    ## First encode the strings to numeric categories
    enc = preprocessing.LabelEncoder()
    enc.fit(cat_feature)
    enc_cat_feature = enc.transform(cat_feature)
    ## Now, apply one hot encoding
    ohe = preprocessing.OneHotEncoder()
    encoded = ohe.fit(enc_cat_feature.reshape(-1,1))
    return encoded.transform(enc_cat_feature.reshape(-1,1)).toarray()
    
categorical_columns= ['economic_typology','year']

for col in categorical_columns:
    temp = encode_string(df[col])
    Features = np.concatenate([Features, temp], axis = 1)

print(Features.shape)
print(Features[:2, :])    

numcols= df.dtypes[df.dtypes != "object"].index

list(numcols)

x = np.concatenate([df[numcols],Features],axis=1)

y=np.array(y)

print(x.shape)
print(y.shape)

# splt train test
nr.seed(9988)
x_train, x_test, y_train, y_test = train_test_split(x,y)

## Train models selected

Boosting_mod = GradientBoostingRegressor (learning_rate=0.1)
Forest_mod = RandomForestRegressor()

Boosting_mod.fit(x_train,y_train)
Forest_mod.fit(x_train,y_train)

print(Boosting_mod.score(x_test,y_test))
print(Forest_mod.score(x_test,y_test))

def print_metrics(y_true, y_predicted, n_parameters):
    ## First compute R^2 and the adjusted R^2
    r2 = sklm.r2_score(y_true, y_predicted)
    r2_adj = r2 - (n_parameters - 1)/(y_true.shape[0] - n_parameters) * (1 - r2)
    
    ## Print the usual metrics and the R^2 values
    print('Mean Square Error      = ' + str(sklm.mean_squared_error(y_true, y_predicted)))
    print('Root Mean Square Error = ' + str(math.sqrt(sklm.mean_squared_error(y_true, y_predicted))))
    print('Mean Absolute Error    = ' + str(sklm.mean_absolute_error(y_true, y_predicted)))
    print('Median Absolute Error  = ' + str(sklm.median_absolute_error(y_true, y_predicted)))
    print('R^2                    = ' + str(r2))
    print('Adjusted R^2           = ' + str(r2_adj))
   

y_Boosting = Boosting_mod.predict(x_test)
y_Forest = Forest_mod.predict(x_test)

   

print_metrics(y_test, y_Boosting, 28)
print_metrics(y_test, y_Forest, 28)

# cross validation
Results_Boosting = cross_validate(Boosting_mod,x,y,scoring="r2",cv=5)
Results_Forest = cross_validate(Forest_mod,x,y,scoring="r2",cv=5)


Boosting_test_scores = Results_Boosting['test_score']
Boosting_train_scores = Results_Boosting['train_score']
Forest_test_scores = Results_Forest['test_score']
Forest_train_scores = Results_Forest['train_score']

print(np.mean(Boosting_train_scores))
print(np.mean(Boosting_test_scores))
print(np.mean(Forest_train_scores))
print(np.mean(Forest_test_scores))

nr.seed(123)
inside = ms.KFold(n_splits=5, shuffle = True)
nr.seed(321)
outside = ms.KFold(n_splits=5, shuffle = True)

nr.seed(3456)
## Define the dictionary for the grid search and the model object to search on
param_grid_Boosting = {"n_estimators": range(20,501,20)}
param_grid_Forest = {"n_estimators": range(20,501,20)}

## Perform the grid search over the parameters
Grid_Boosting = ms.GridSearchCV(estimator = Boosting_mod, param_grid = param_grid_Boosting, 
                      cv = inside, # Use the inside folds
                      scoring = 'r2',
                      return_train_score = True)

Grid_Forest = ms.GridSearchCV(estimator = Forest_mod, param_grid = param_grid_Forest, 
                      cv = inside, # Use the inside folds
                      scoring = 'r2',
                      return_train_score = True)


## Fit thhe cross validated grid search over the data 
Grid_Boosting.fit(x,y)
Grid_Forest.fit(x,y)



 Grid_Boosting.best_params_, Grid_Boosting.best_score_

Grid_Forest.best_params_, Grid_Forest.best_score_

final_results_Boosting = cross_validate(Grid_Boosting.best_estimator_,x_train,y_train)
final_results_Forest = cross_validate(Grid_Forest.best_estimator_,x_train,y_train)

test_scores_Boosting = final_results_Boosting['test_score']
train_scores_Boosting = final_results_Boosting['train_score']
print(np.mean(train_scores_Boosting))
print(np.mean(test_scores_Boosting))

test_scores_Forest = final_results_Forest['test_score']
train_scores_Forest = final_results_Forest['train_score']
print(np.mean(train_scores_Forest))
print(np.mean(test_scores_Forest))

estimator = GradientBoostingRegressor(learning_rate=0.1,
                                      n_estimators=500)
estimator.fit(x_train,y_train)

estimator.score(x_test,y_test)

estimator2= RandomForestRegressor(n_estimators=440)
estimator2.fit(x_train,y_train)

estimator2.score(x_test,y_test)

### prediction

test_values_processed=np.array(pd.read_csv('test_values_processed.csv', header=None))


estimator2.predict(test_values_processed)

prediccion=pd.DataFrame(estimator2.predict(test_values_processed),columns=['evictions'])
prediccion.index.names=['row_id']

prediccion['evictions']=np.expm1(prediccion['evictions'])

prediccion['evictions']= prediccion['evictions'].astype(np.int64)
prediccion

prediccion.to_csv('prediccion.csv')
