==> MPP_DS/4 Auto ML TPOT.ipynb <==
#Fundamental librarys to math and stats process
import numpy as np
import numpy.random as nr
import scipy.stats as ss
import math
#data prepared
import pandas as pd

#ML preprocessi
from sklearn import preprocessing
from sklearn.preprocessing import StandardScaler
import sklearn.model_selection as ms
from sklearn.pipeline import make_pipeline
from sklearn import feature_selection as fs

# ML algorithms models
from sklearn.ensemble import RandomForestRegressor,GradientBoostingRegressor 
from sklearn import linear_model as lm
import xgboost as xgb

# ML Evaluations
import sklearn.metrics as sklm
from sklearn import metrics
from sklearn.model_selection import cross_validate, train_test_split

#Auto ML
from tpot import TPOTRegressor

#Ploting
import matplotlib.pyplot as plt
import seaborn as sns


%matplotlib inline

import warnings
warnings.filterwarnings('ignore')

df_trf = (pd.read_csv('df_trf.csv'))
df_enc = (pd.read_csv('df_enc.csv'))
print(df_trf.columns)
print(df_enc.columns)



x_trf= df_trf.drop(['Unnamed: 0','population','evictions'], axis=1)
y_trf= df_trf['evictions']
x_trf = np.array(x_trf)
y_trf=np.array(y_trf)

x_enc= df_enc.drop(['Unnamed: 0','population','evictions'], axis=1)
y_enc= df_enc['evictions']
x_enc = np.array(x_enc)
y_enc=np.array(y_enc)

print(x_enc.shape)
print(y_enc.shape)

print(x_trf.shape)
print(y_trf.shape)

# splt train test
nr.seed(9988)
xtrf_train, xtrf_test, ytrf_train, ytrf_test = train_test_split(x_trf,y_trf,test_size=0.1)
xenc_train, xenc_test, yenc_train, yenc_test = train_test_split(x_enc,y_enc,test_size=0.1)

## Train models Auto ML

tpot_trf = TPOTRegressor(generations=7, population_size=30, cv=5,verbosity=2, scoring='r2')
tpot_trf.fit(xtrf_train, ytrf_train)
tpot_trf.export('tpot_pipeline_trf.py')

tpot_enc = TPOTRegressor(generations=7, population_size=30, cv=5,verbosity=2, scoring='r2')
tpot_enc.fit(xenc_train, yenc_train)
tpot_enc.export('tpot_pipeline_enc.py')



# trf pipeline

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.pipeline import make_pipeline

from sklearn.preprocessing import Normalizer, MinMaxScaler,FunctionTransformer, PolynomialFeatures
from copy import copy

from tpot.builtins import StackingEstimator
from sklearn.ensemble import RandomForestRegressor,ExtraTreesRegressor
from sklearn.neighbors import KNeighborsRegressor




pipeline_trf1 = make_pipeline(
    Normalizer(norm="l2"),
    RandomForestRegressor(bootstrap=False,
                          max_features=0.7000000000000001,
                          min_samples_leaf=3,
                          min_samples_split=2,
                          n_estimators=100)
)

# Average CV score on the training set was:0.8812547047881732

pipeline_trf2 = make_pipeline(
    MinMaxScaler(),
    RandomForestRegressor(bootstrap=False,
                          max_features=0.7000000000000001,
                          min_samples_leaf=6,
                          min_samples_split=2,
                          n_estimators=100))

# Average CV score on the training set was:0.8925027080134283# 
pipeline_trf3 = make_pipeline(
    make_union( FunctionTransformer(copy),FunctionTransformer(copy)),
    ExtraTreesRegressor(
        bootstrap=False,
        max_features=0.7500000000000001,
        min_samples_leaf=3,
        min_samples_split=3, 
        n_estimators=100)
)


#Average CV score on the training set was:0.9021060044881622
pipeline_trf4 = make_pipeline(
    MinMaxScaler(),
    PolynomialFeatures(degree=2, include_bias=False, interaction_only=False),
    StackingEstimator(
        estimator=KNeighborsRegressor(
            n_neighbors=34, p=1, weights="uniform")),
    StackingEstimator(
        estimator=ExtraTreesRegressor(
            bootstrap=False,
            max_features=0.45,
            min_samples_leaf=5,
            min_samples_split=15,
            n_estimators=100)),
    ExtraTreesRegressor(
        bootstrap=False,
        max_features=0.2,
        min_samples_leaf=1,
        min_samples_split=3,
        n_estimators=100)
)


# enc pipeline
import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.pipeline import make_pipeline, make_union
from tpot.builtins import StackingEstimator,OneHotEncoder

from sklearn.linear_model import RidgeCV, ElasticNetCV
from sklearn.ensemble import RandomForestRegressor
from xgboost import XGBRegressor


# Average CV score on the training set was:0.8767668328812122
pipeline_enc1 = make_pipeline(
    StackingEstimator(estimator=ElasticNetCV(l1_ratio=0.9500000000000001, tol=1e-05)),
    RandomForestRegressor(bootstrap=False, max_features=0.1, min_samples_leaf=1, min_samples_split=4, n_estimators=100)
)

# Average CV score on the training set was:0.8630661690341384

pipeline_enc2 = RandomForestRegressor(bootstrap=False,
                                      max_features=0.25, 
                                      min_samples_leaf=2,
                                      min_samples_split=2, 
                                      n_estimators=100)

# Average CV score on the training set was:0.8409039139854155
pipeline_enc3 = make_pipeline(
    OneHotEncoder(minimum_fraction=0.1, sparse=False, threshold=10),
    StackingEstimator(
        estimator=XGBRegressor(
            learning_rate=0.01,
            max_depth=5,
            min_child_weight=1,
            n_estimators=100, 
            nthread=1,
            subsample=0.6000000000000001)),
    RidgeCV()
)

# Average CV score on the training set was:0.8784382472941576

pipeline_enc4 = make_pipeline(
    StackingEstimator(estimator=RandomForestRegressor(bootstrap=False, max_features=0.5, min_samples_leaf=2, min_samples_split=4, n_estimators=100)),
    XGBRegressor(learning_rate=1.0, max_depth=5, min_child_weight=12, n_estimators=100, nthread=1, subsample=0.8500000000000001)
)

pipeline_trf1.fit(xtrf_train,ytrf_train)
pipeline_trf2.fit(xtrf_train,ytrf_train)
pipeline_trf3.fit(xtrf_train,ytrf_train)
pipeline_trf4.fit(xtrf_train,ytrf_train)

pipeline_enc1.fit(xenc_train,yenc_train)
pipeline_enc2.fit(xenc_train,yenc_train)
pipeline_enc3.fit(xenc_train,yenc_train)
pipeline_enc4.fit(xenc_train,yenc_train)

#cross validation
Results_pipeline_trf1 = cross_validate(pipeline_trf1,x_trf,y_trf,scoring="r2",cv=5)
Results_pipeline_trf2 = cross_validate(pipeline_trf2,x_trf,y_trf,scoring="r2",cv=5)
Results_pipeline_trf3 = cross_validate(pipeline_trf3,x_trf,y_trf,scoring="r2",cv=5)
Results_pipeline_trf4 = cross_validate(pipeline_trf4,x_trf,y_trf,scoring="r2",cv=5)

Results_pipeline_enc1 = cross_validate(pipeline_enc1 ,x_enc,y_enc,scoring="r2",cv=5)
Results_pipeline_enc2 = cross_validate(pipeline_enc2 ,x_enc,y_enc,scoring="r2",cv=5)
Results_pipeline_enc3 = cross_validate(pipeline_enc3 ,x_enc,y_enc,scoring="r2",cv=5)
Results_pipeline_enc4 = cross_validate(pipeline_enc1 ,x_enc,y_enc,scoring="r2",cv=5)



pipeline_trf1_test_scores = Results_pipeline_trf1['test_score']
pipeline_trf1_train_scores = Results_pipeline_trf1['train_score']

pipeline_trf2_test_scores = Results_pipeline_trf2['test_score']
pipeline_trf2_train_scores = Results_pipeline_trf2['train_score']

pipeline_trf3_test_scores = Results_pipeline_trf3['test_score']
pipeline_trf3_train_scores = Results_pipeline_trf3['train_score']

pipeline_trf4_test_scores = Results_pipeline_trf4['test_score']
pipeline_trf4_train_scores = Results_pipeline_trf4['train_score']



print('trf1 train='+str(np.mean(pipeline_trf1_train_scores)))
print('trf1 test='+str(np.mean(pipeline_trf1_test_scores)))

print('trf2 train='+str(np.mean(pipeline_trf2_train_scores)))
print('trf2 test='+str(np.mean(pipeline_trf2_test_scores)))

print('trf3 train='+str(np.mean(pipeline_trf3_train_scores)))
print('trf3 test='+str(np.mean(pipeline_trf3_test_scores)))


print('trf4 train='+str(np.mean(pipeline_trf4_train_scores)))
print('trf4 test='+str(np.mean(pipeline_trf4_test_scores)))


pipeline_enc1_test_scores = Results_pipeline_enc1['test_score']
pipeline_enc1_train_scores = Results_pipeline_enc1['train_score']

pipeline_enc2_test_scores = Results_pipeline_enc2['test_score']
pipeline_enc2_train_scores = Results_pipeline_enc2['train_score']

pipeline_enc3_test_scores = Results_pipeline_enc3['test_score']
pipeline_enc3_train_scores = Results_pipeline_enc3['train_score']

pipeline_enc4_test_scores = Results_pipeline_enc4['test_score']
pipeline_enc4_train_scores = Results_pipeline_enc4['train_score']



print('enc1 train='+str(np.mean(pipeline_enc1_train_scores)))
print('enc1 test='+str(np.mean(pipeline_enc1_test_scores)))

print('enc2 train='+str(np.mean(pipeline_enc2_train_scores)))
print('enc2 test='+str(np.mean(pipeline_enc2_test_scores)))

print('enc3 train='+str(np.mean(pipeline_enc3_train_scores)))
print('enc3 test='+str(np.mean(pipeline_enc3_test_scores)))

print('enc4 train='+str(np.mean(pipeline_enc4_train_scores)))
print('enc4 test='+str(np.mean(pipeline_enc4_test_scores)))

pipeline_trf=pipeline_trf4.fit(xtrf_train,ytrf_train)
pipeline_enc=pipeline_enc1.fit(xenc_train,yenc_train)

print(pipeline_trf4.score(xtrf_test,ytrf_test))
print(pipeline_enc1.score(xenc_test,yenc_test))



### Preidction 

test_trf=pd.read_csv('df_test_trf.csv')
test_trf=test_trf.drop(['Unnamed: 0','median_household_income','median_property_value'],axis=1)
test_trf= np.array(test_trf)

test_enc=pd.read_csv('df_test_enc.csv')
test_enc=test_enc.drop(['Unnamed: 0','median_household_income','median_property_value'],axis=1)
test_enc= np.array(test_trf)


AML_Prediction=pipeline_trf.predict(test_trf)
AML_Prediction= pd.DataFrame(AML_Prediction,columns=['evictions'])
AML_Prediction.index.names=['row_id']
AML_Prediction['evictions']=np.expm1(AML_Prediction['evictions'])
AML_Prediction['evictions']=AML_Prediction['evictions'].astype(np.int64)
AML_Prediction.head()



AML_Prediction.to_csv('AML_Prediction_trf1.csv')


AML_Prediction2=pipeline_enc.predict(test_enc)
AML_Prediction2= pd.DataFrame(AML_Prediction2,columns=['evictions'])
AML_Prediction2.index.names=['row_id']
AML_Prediction2['evictions']=AML_Prediction2['evictions'].astype(np.int64)
AML_Prediction2.head()

AML_Prediction2.to_csv('AML_Prediction_Enc1.csv')
